# [ASStackLayoutSpec] Performance improvements