Moves layout calculations of the stack layout spec in the C++ layer of the layout spec implementation. Previously we used references of `ASLayout` to as for specific properties like `layoutElement` and `style` or `size` on the `ASLlayoutElement`.

Now we create a C++ wrapper around the `ASLayout` before the actual layout calculations happening and use this to access the common components to reduce the overhead of Objc message calling.

- Add `ASStackLayoutSpecItem ` and don't separate between unpositioned and positioned items anymore. This didn't work properly previously in the first place as the `ASLayout` that was included was basically shared, so on ASDK side having them was basically nonsense. It would have make sense if we would have create specific `ASLayout` objects for unpositioned and positioned items by copying the old one. In that case the immutability of the `ASLayout` would have make sense.
- Reduce overhead of calling `.style` `.style.size` on `ASLayoutElement`s
- Bypass baseline calculation if not needed