@maicki 2016-10-25
@rcancro Hey Ricky, could you please give that a first review pass. Furthermore do know of any other case where we could bypass the baseline pass? Thanks!