@appleguy 2016-10-29
This does appear to be a significant performance improvement — Great work @maicki !!  This makes a big difference on not only our benchmarks, but also real-world app workloads!

<img width="1240" alt="screen shot 2016-10-28 at 8 32 41 pm" src="https://cloud.githubusercontent.com/assets/565251/19826955/1da22682-9d4e-11e6-9357-fcb6c02fb2bd.png">
<img width="1222" alt="screen shot 2016-10-28 at 8 30 13 pm" src="https://cloud.githubusercontent.com/assets/565251/19826956/1dbb042c-9d4e-11e6-8c31-a8c34cfa19e3.png">