@appleguy 2016-10-30
Benchmark went from:

Old, ASDK: 1.46 ops/sec (100 iterations - worst case for ASDK perf)
New, ASDK: 2.26 ops/sec (100 iterations)
Autolayout: 0.26 ops/sec (100 iterations)
LayoutKit: 2.56 ops/sec (100 iterations)

ASDK has substantial perf gains over LayoutKit in all of the more normal use cases, especially because it does not actually wait on / flush all of the layout at once as this benchmark tests.

In the new version, here are some optimization targets:

<img width="557" alt="screen shot 2016-10-29 at 6 10 43 pm" src="https://cloud.githubusercontent.com/assets/565251/19833832/7fd500a6-9e05-11e6-9f54-30c642e51b7f.png">
<img width="939" alt="screen shot 2016-10-29 at 6 06 08 pm" src="https://cloud.githubusercontent.com/assets/565251/19833833/7ff19266-9e05-11e6-98d9-cbea40b9c707.png">
<img width="1038" alt="screen shot 2016-10-29 at 6 05 27 pm" src="https://cloud.githubusercontent.com/assets/565251/19833834/7ff19874-9e05-11e6-8863-fb32313d7434.png">