@maicki 2016-10-30
@rcancro @appleguy Thanks for reviewing! I will let the variable name for now as it is as we use this name all in the stack layout spec components from the beginning on. We could address it in a follow up PR, it would add some noise that is not really related to the task to this PR.

Will merge as soon as CI finished.